server/pyproject.toml
[tool.poetry]
name = "jsa-demo-app"
version = "0.0.1"
description = "This application showcases the main features of JSAlchemy"
authors = ["Giuseppe Di Bona <giusedb@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy-js-authentication = "^0.1.0"
sqlalchemy-js-api = "^0.1.0"
sqlalchemy-js-authorization = "^0.1.0"
click = "^8.1.7"
pytest = "^8.3.3"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
alembic = "^1.14.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
pyyaml = "^6.0.2"
aiosqlite = "^0.20.0"
trio = "^0.27.0"
asyncpg = "^0.30.0"
aiohttp = "^3.11.11"
pytest-mock = "^3.14.0"
blacksheep = {extras = ["full"], version = "^2.0.8"}
pluralizer = "^1.2.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
orjson = "^3.11.1"
fakeredis = "^2.32.0"
aiofile = "^3.9.0"

[tool.poetry.group.local.dependencies]
sqlalchemy-js-authentication = { path = "../libs/jsa-authentication", develop = true }
sqlalchemy-js-authorization = { path = "../libs/jsa-authorization", develop = true }
sqlalchemy-js-api = { path = "../libs/jsa-api", develop = true }
sqlalchemy-js-web-context = { path = "../libs/jsa-context", develop = true }

[tool.poetry.group.dev.dependencies]
ipython = "^8.20.0"
pylint = "^3.3.1"
requests = "^2.32.3"
python-lsp-server = {extras = ["all"], version = "^1.12.0"}


[tool.poetry.group.perhaps.dependencies]
wrapt = "^1.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
test = "pytest"
tapp.script = "cli:main_cli"
